feat: 타임어택, 서바이버 시간초 소숫점 셋째자리까지 계산하게 로직 수정

+++ backend/src/main/java/com/problemio/challenge/domain/ChallengeRanking.java
@@ -14,6 +14,6 @@ public class ChallengeRanking {
     private Long userId;
     private Integer ranking;
     private Double score;
-    private Integer playTime;
+    private Double playTime;
     private LocalDateTime createdAt;
 }
